+++ src/Exceptions/GuardMismatch.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Spatie\Permission\Exceptions;
-
-use Exception;
-
-class GuardMismatch extends Exception
-{
-}