+++ src/Models/Permission.php
@@ -27,7 +27,7 @@ class Permission extends Model implements PermissionContract
 
         parent::__construct($attributes);
 
-        $this->setTable(config('laravel-permission.table_names.permissions'));
+        $this->setTable(config('permission.table_names.permissions'));
     }
 
     /**
@@ -36,8 +36,8 @@ class Permission extends Model implements PermissionContract
     public function roles(): BelongsToMany
     {
         return $this->belongsToMany(
-            config('laravel-permission.models.role'),
-            config('laravel-permission.table_names.role_has_permissions')
+            config('permission.models.role'),
+            config('permission.table_names.role_has_permissions')
         );
     }